requirements.txt
androguard==3.3.5
anytree==2.8.0
apk-parse==1.0.0
apkutils==0.6.6
asgiref==3.2.3
asn1crypto==1.3.0
awscli==1.18.76
backcall==0.1.0
beautifulsoup4==4.8.2
blinker==1.4
boto3==1.13.26
botocore==1.16.26
Brotli==1.0.7
bytecode==0.9.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cigam==0.0.3
Click==7.0
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.1
dj-database-url==0.5.0
Django==3.0.1
django-filter==2.3.0
django-heroku==0.3.1
django-mysql==3.3.0
django-storages==1.9.1
django-tables2==2.3.1
dnspython==1.16.0
docutils==0.15.2
Flask==1.1.2
future==0.18.2
googleplay-api==0.1.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.8
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.2
jmespath==0.10.0
kaitaistruct==0.8
kiwisolver==1.1.0
ldap3==2.7
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mitmproxy==5.1.1
mysql-connector-python==8.0.19
networkx==2.4
numpy==1.18.1
pandas==1.0.1
parse==1.14.0
parso==0.5.2
passlib==1.7.2
pickleshare==0.7.5
Pillow==7.1.2
play-store==1.0
prompt-toolkit==3.0.2
protobuf==3.6.1
psycopg2==2.8.4
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.19
pydivert==2.1.0
pydot==1.4.1
pyelftools==0.26
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyperclip==1.8.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
rsa==3.4.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.14.0
sortedcontainers==2.1.0
soupsieve==1.9.5
sqlparse==0.3.0
tabulate==0.8.7
TextWizard==0.0.3
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.8
urwid==2.1.0
utils==1.0.0
wcwidth==0.1.8
Werkzeug==1.0.1
whitenoise==5.1.0
wsproto==0.15.0
xmltodict==0.12.0
zstandard==0.13.0
